refactor: move domain validation into validate_endpoint.go
Consolidate validateTokenEndpointDomain and validateAccountApiUrlDomain
into a single validateDynatraceDomain helper in validate_endpoint.go.
Both functions checked that a URL hostname is on the Dynatrace domain
allowlist to prevent SSRF and credential redirection attacks.

- Add validate_endpoint.go with shared validateDynatraceDomain helper
  - Case-insensitive host matching via strings.ToLower (RFC-safe)
  - Accepts fieldName param for clear error messages
- Add validate_endpoint_test.go with all tokenEndpoint + apiUrl cases
- Inline calls in parseOAuth() and parseSingleAccount(); remove wrappers
- Remove now-unused net/url and strings imports from account.go
