+++ pkg/manifest/loader/account_test.go
@@ -337,103 +337,3 @@ func deepCopy(t *testing.T, in persistence.Account) persistence.Account {
 	assert.NoError(t, e)
 	return o
 }
-func TestValidateAccountApiUrlDomain(t *testing.T) {
-t.Run("accepts valid dynatrace.com host", func(t *testing.T) {
-assert.NoError(t, validateAccountApiUrlDomain("https://api.dynatrace.com"))
-})
-
-t.Run("accepts valid dynatracelabs.com host", func(t *testing.T) {
-assert.NoError(t, validateAccountApiUrlDomain("https://sso.dynatracelabs.com"))
-})
-
-t.Run("accepts subdomain on allowed domain", func(t *testing.T) {
-assert.NoError(t, validateAccountApiUrlDomain("https://abc.def.dynatrace.com"))
-})
-
-t.Run("accepts uppercase host (case-insensitive)", func(t *testing.T) {
-assert.NoError(t, validateAccountApiUrlDomain("https://API.DYNATRACE.COM"))
-})
-
-t.Run("accepts host with port", func(t *testing.T) {
-assert.NoError(t, validateAccountApiUrlDomain("https://api.dynatrace.com:8443/account"))
-})
-
-t.Run("rejects attacker-controlled host", func(t *testing.T) {
-err := validateAccountApiUrlDomain("https://attacker.example.com/steal")
-require.Error(t, err)
-assert.Contains(t, err.Error(), "not allowed")
-})
-
-t.Run("rejects look-alike suffix without leading dot", func(t *testing.T) {
-err := validateAccountApiUrlDomain("https://evil-dynatrace.com")
-require.Error(t, err)
-assert.Contains(t, err.Error(), "not allowed")
-})
-
-t.Run("rejects domain that only ends in dynatrace.com via path", func(t *testing.T) {
-err := validateAccountApiUrlDomain("https://attacker.com/api.dynatrace.com")
-require.Error(t, err)
-})
-
-t.Run("rejects RFC1918 SSRF target", func(t *testing.T) {
-err := validateAccountApiUrlDomain("http://10.0.0.1:8080/admin")
-require.Error(t, err)
-})
-
-t.Run("rejects cloud metadata SSRF target", func(t *testing.T) {
-err := validateAccountApiUrlDomain("http://169.254.169.254/latest/meta-data/")
-require.Error(t, err)
-})
-
-t.Run("rejects userinfo prefix masquerade", func(t *testing.T) {
-// https://api.dynatrace.com@attacker.com -> actual host is attacker.com
-err := validateAccountApiUrlDomain("https://api.dynatrace.com@attacker.com/")
-require.Error(t, err)
-})
-
-t.Run("rejects malformed URL", func(t *testing.T) {
-err := validateAccountApiUrlDomain("::::not a url")
-require.Error(t, err)
-assert.Contains(t, err.Error(), "not a valid URL")
-})
-}
-
-func TestParseSingleAccount_RejectsNonDynatraceApiUrl(t *testing.T) {
-ctx := &Context{Opts: Options{}}
-a := persistence.Account{
-Name:        "name",
-AccountUUID: persistence.TypedValue{Value: "8f9935ee-2068-455d-85ce-47447f19d5d5"},
-ApiUrl:      &persistence.TypedValue{Value: "https://attacker.example.com/steal"},
-OAuth: persistence.OAuth{
-ClientID:     persistence.AuthSecret{Name: "ID"},
-ClientSecret: persistence.AuthSecret{Name: "SECRET"},
-},
-}
-t.Setenv("ID", "id")
-t.Setenv("SECRET", "secret")
-
-_, err := parseSingleAccount(ctx, a)
-require.Error(t, err)
-assert.Contains(t, err.Error(), "apiUrl")
-assert.Contains(t, err.Error(), "not allowed")
-}
-
-func TestParseSingleAccount_AcceptsDynatraceApiUrl(t *testing.T) {
-ctx := &Context{Opts: Options{}}
-a := persistence.Account{
-Name:        "name",
-AccountUUID: persistence.TypedValue{Value: "8f9935ee-2068-455d-85ce-47447f19d5d5"},
-ApiUrl:      &persistence.TypedValue{Value: "https://api.dynatrace.com"},
-OAuth: persistence.OAuth{
-ClientID:     persistence.AuthSecret{Name: "ID"},
-ClientSecret: persistence.AuthSecret{Name: "SECRET"},
-},
-}
-t.Setenv("ID", "id")
-t.Setenv("SECRET", "secret")
-
-acc, err := parseSingleAccount(ctx, a)
-require.NoError(t, err)
-require.NotNil(t, acc.ApiUrl)
-assert.Equal(t, "https://api.dynatrace.com", acc.ApiUrl.Value)
-}
